update page layout, move auth pages to auth folder, fix redirect for users with isAdmin role, fix bugs
