+++ src/app/(auth)/partner-register/page.jsx
