+++ src/app/clients/search/[id]/page.jsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { PageTitle } from '@/components/typography/PageTitle'
+import { PageContainer, PageTitle } from '@/components/typography/PageTitle'
 import useUserStore from '@/store/zustand'
 import { TableClientDetails } from './table-client-details'
 import { TableOrderItems } from './table-order-items'
@@ -12,12 +12,12 @@ export default function Page({ params }) {
   const client = clients && clients?.find((client) => client.id === params.id)
 
   return (
-    <div className="py-10 pl-5 pr-2">
+    <PageContainer>
       <PageTitle>Customer Information</PageTitle>
       {client && <TableClientDetails client={client} />}
 
       <PageTitle>Order History</PageTitle>
       <TableOrderItems clientId={client?.id} />
-    </div>
+    </PageContainer>
   )
 }
