+++ src/app/clients/search/page.jsx
@@ -1,16 +1,16 @@
 'use client'
 
-import { protectRoute } from '@/app/auth-components/protect-route'
-import { PageTitle } from '@/components/typography/PageTitle'
+import { protectRoute } from '@/app/(auth)/_components/protect-route'
+import { PageContainer, PageTitle } from '@/components/typography/PageTitle'
 import { TableClients } from './table-clients'
 
 function Page() {
   return (
-    <div className="py-10 pl-5 pr-2">
+    <PageContainer>
       <PageTitle>Search Customer</PageTitle>
       <TableClients />
-    </div>
+    </PageContainer>
   )
 }
 
-export default protectRoute(Page)
+export default protectRoute(Page, 'clients/search')
