+++ src/app/messages/productquery/page.jsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { protectRoute } from '@/app/auth-components/protect-route'
-import { PageTitle } from '@/components/typography/PageTitle'
+import { protectRoute } from '@/app/(auth)/_components/protect-route'
+import { PageContainer, PageTitle } from '@/components/typography/PageTitle'
 import { Button } from '@/components/ui/button'
 import { useRouter } from 'next/navigation'
 import React from 'react'
@@ -25,14 +25,14 @@ function Page() {
   )
 
   return (
-    <div className="py-10 pl-5 pr-2">
+    <PageContainer>
       <PageTitle>
         <BackButton />
         고객 문의
       </PageTitle>
       <TableProductQuery />
-    </div>
+    </PageContainer>
   )
 }
 
-export default protectRoute(Page)
+export default protectRoute(Page, 'productquery')
