+++ src/app/messages/productquery/table-product-query.jsx
@@ -2,11 +2,12 @@
 
 import { useFetchAuth } from '@/app/orders/manage/use-fetch-auth'
 import { DataTable } from '@/components/data-table/data-table'
+import { useSession } from 'next-auth/react'
 import React from 'react'
 import { columns } from './columns'
 
 export function TableProductQuery() {
-  const { session } = useSession()
+  const { data: session } = useSession()
   const sellerId = session?.user?._id
   const url = sellerId ? `questions/vendor/${sellerId}` : null
   const { data, isLoading } = useFetchAuth(url)
