+++ src/app/products/productregister/page.jsx
@@ -21,7 +21,7 @@ import styles from './productregister.module.css'
 import 'react-datetime-picker/dist/DateTimePicker.css'
 import 'react-calendar/dist/Calendar.css'
 import 'react-clock/dist/Clock.css'
-import { protectRoute } from '@/app/auth-components/protect-route'
+import { protectRoute } from '@/app/(auth)/_components/protect-route'
 import baseURL from '@/assets/common/baseUrl'
 import LoadingSpinner from '@/components/LoadingSpinner'
 import axios from 'axios'
@@ -1552,4 +1552,4 @@ function Page() {
   )
 }
 
-export default protectRoute(Page)
+export default protectRoute(Page, 'productregister')
